Spider Crush Android App
The project is about developing 2D video android game application.In
this game,you can smash the spider and score points.Spiders randomly
move from the top of the screen to the bottom of the screen. For each
Spider that reaches the bottom of the screen, one “life” will be lost.